add mapping for ag_cap and ag_cap_source; small refactoring of app_prep etl
