+++ bin/migrate-oats-data/common/oats_decision_enum.py
@@ -1,4 +1,5 @@
 from enum import Enum
+from .alcs_application_enum import AlcsAgCapSource, AlcsAgCap
 
 
 class AlcsDecisionOutcomeCodeEnum(Enum):
@@ -11,3 +12,17 @@ class AlcsNoiModificationOutcomeCodeEnum(Enum):
     PENDING = "PEN"
     PROCEED_TO_MODIFY = "PRC"
     REFUSE_TO_MODIFY = "REF"
+
+
+class OatsToAlcsAgCapSource(Enum):
+    BCLI = AlcsAgCapSource.BCLI.value
+    CLI = AlcsAgCapSource.CLI.value
+    ONSI = AlcsAgCapSource.On_site.value
+
+
+class OatsToAlcsAgCap(Enum):
+    P = AlcsAgCap.Prime.value
+    PD = AlcsAgCap.Prime_Dominant.value
+    MIX = AlcsAgCap.Mixed_Prime_Secondary.value
+    S = AlcsAgCap.Secondary.value
+    U = AlcsAgCap.Unclassified.value
